requirements.txt
fastapi
uvicorn[standard]
google-auth
google-auth-oauthlib
google-api-python-client
openai
python-dotenv
httpx
sqlalchemy
alembic
apscheduler
beautifulsoup4
html2text
